refactor(#236): Integer time estimates
- More in line with how time is kept on Linux

- Should be further refactored to use std::chrono if the task allocation stuff
  ever makes it out of simulation and onto real robots.

- Also re-run formatter.

+++ include/rcppsw/ta/time_estimate.hpp
@@ -34,7 +34,7 @@ NS_START(rcppsw, ta);
 /*******************************************************************************
  * Class Definitions
  ******************************************************************************/
-using time_estimate = math::ema<double>;
+using time_estimate = math::ema<int>;
 
 NS_END(ta, rcppsw);
